refactor: InputController class가 controller의 성격에 맞지 않는 문제 해결
- RacingController class가 InputController class를 제어하지 않도록 수정
